# backend/api-gateway/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Conversor de moedas BRL",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "description": "Login by email and password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@user.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout authenticated user",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Retrieve profile information",
                "description": "Get profile short information",
                "operationId": "profileShow",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/signup": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Signup",
                "description": "Signup",
                "operationId": "authSignup",
                "requestBody": {
                    "description": "Pass user data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Jhon"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@user.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/currency": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Retrieve available currencies",
                "operationId": "currencyIndex",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/exchange": {
            "get": {
                "tags": [
                    "Exchanges"
                ],
                "summary": "Retrieve exchanges created by current authenticated user",
                "operationId": "exchangeIndex",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Exchanges"
                ],
                "summary": "Create a new exchange",
                "operationId": "exchangeCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "origin_value",
                                    "destination_currency_id",
                                    "payment_method_id"
                                ],
                                "properties": {
                                    "origin_value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "1000"
                                    },
                                    "destination_currency_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "payment_method_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/exchange/{id}": {
            "get": {
                "tags": [
                    "Exchanges"
                ],
                "summary": "Retrieve an exchange by id",
                "operationId": "exchangeShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Exchange id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/fee": {
            "get": {
                "tags": [
                    "Exchange fees"
                ],
                "summary": "Retrieve fees",
                "operationId": "feeIndex",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Exchange fees"
                ],
                "summary": "Create fee",
                "operationId": "feeCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fee_rate",
                                    "starting_value"
                                ],
                                "properties": {
                                    "starting_value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "1000"
                                    },
                                    "fee_rate": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "1.5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/fee/{id}": {
            "get": {
                "tags": [
                    "Exchange fees"
                ],
                "summary": "Retrieve fee by id",
                "operationId": "feeShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Fee id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Exchange fees"
                ],
                "summary": "Update fee by id",
                "operationId": "feeUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Fee id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fee_rate",
                                    "starting_value"
                                ],
                                "properties": {
                                    "starting_value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "1000"
                                    },
                                    "fee_rate": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "1.5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Exchange fees"
                ],
                "summary": "Delete fee by id",
                "operationId": "feeDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Fee id",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payment_method": {
            "get": {
                "tags": [
                    "Payment methods"
                ],
                "summary": "Retrieve payment methods",
                "operationId": "paymentMethodIndex",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payment_method/{id}": {
            "get": {
                "tags": [
                    "Payment methods"
                ],
                "summary": "Retrieve payment method by id",
                "operationId": "paymentMethodShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment method id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Payment methods"
                ],
                "summary": "Update payment method by id",
                "operationId": "paymentMethodUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment method id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fee_rate"
                                ],
                                "properties": {
                                    "fee_rate": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "1.5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}